updates from Richelle emaill 12/19/16
The POI map is finally approved, so we can upload that onto the site
(it’s in the drive). It goes right under the “check out the
neighborhood” header (above the fillable form).
Done
Under floor plans, can we increase the padding on the top and bottom of
the plans? Let’s try 5px (?)
Done
and can we also do that on the site plan image under “community”.
Padding on all 4 sides please!
Done
The gallery could use some on the top and bottom too.
Done?  I added 5px to the top and bottom of the main pic (not sure what
this one is getting at...)
There is also an updated site plan image - I’ve uploaded that onto the
drive as well.
Added